Cast the result of ~ back to the part type
~v promotes uint8_t to int, so assigning it back narrows. Clang rejects
that under -Wimplicit-int-conversion, which kept these tests from
building in projects that compile ass with the shared warning set.

+++ modules/AssTests/code/private/bit_span_tests.cpp
@@ -278,7 +278,7 @@ TEST(BitSpanTest, CountOnes)
 
             for (auto& v : buffer)
             {
-                v = ~v;
+                v = static_cast<uint8_t>(~v);
             }
 
             ASSERT_EQ(bit_span.CountOnes(), odd_count);
